+++ src/poly/mod.rs
@@ -1,2 +1,4 @@
+pub mod domain;
+pub mod evaluations;
 /// Modules for working with polynomials in coefficient forms.
 pub mod polynomial;